Just some small fixes
This is my first contribution BTW.

What did I do?

I've added translations to stuff like the toolgun, the large ion thruster blocks and the cannon.

I've also updated the dependencies of the mod and I've added an implementation method for zerocore and cc-tweaked (because it would crash the game without this change.)

And lastly I've added myself to the contributors.md file.

I know it isn't a lot, but I will hopefully get around to doing some more stuff for this mod.